+++ cmd/goat/output/some/some.go
@@ -12,7 +12,9 @@ package some
 
 import (
 	"fmt"
+	"net/netip"
 	"slices"
+	"time"
 
 	"github.com/robert-kisteleki/goat"
 	"github.com/robert-kisteleki/goat/cmd/goat/output"
@@ -21,6 +23,8 @@ import (
 
 var verbose bool
 var total uint
+var connectLastResults map[uint]*result.ConnectionResult
+var connectTableOutput bool
 
 func init() {
 	output.Register("some", supports, setup, start, process, finish)
@@ -38,48 +42,73 @@ func supports(outtype string) bool {
 
 func setup(isverbose bool, options []string) {
 	verbose = isverbose
+	for _, opt := range options {
+		if opt == "table" {
+			connectTableOutput = true
+			break
+		}
+	}
 }
 
 func start() {
+	connectLastResults = make(map[uint]*result.ConnectionResult)
 }
 
 func process(res any) {
-	total++
-
+	out := ""
 	switch t := res.(type) {
 	case *result.Result:
 		switch rt := (*t).(type) {
 		case *result.PingResult:
-			fmt.Println(SomeOutputPing(rt))
+			out = SomeOutputPing(rt)
 		case *result.DnsResult:
-			fmt.Println(SomeOutputDns(rt))
+			out = SomeOutputDns(rt)
 		case *result.TracerouteResult:
-			fmt.Println(SomeOutputTraceroute(rt))
+			out = SomeOutputTraceroute(rt)
 		case *result.CertResult:
-			fmt.Println(SomeOutputCert(rt))
+			out = SomeOutputCert(rt)
 		case *result.HttpResult:
-			fmt.Println(SomeOutputHttp(rt))
+			out = SomeOutputHttp(rt)
 		case *result.NtpResult:
-			fmt.Println(SomeOutputNtp(rt))
+			out = SomeOutputNtp(rt)
 		case *result.ConnectionResult:
-			fmt.Println(SomeOutputConnection(rt))
+			out = SomeOutputConnection(rt)
 		case *result.UptimeResult:
-			fmt.Println(SomeOutputUptime(rt))
+			out = SomeOutputUptime(rt)
 		}
 	case goat.AsyncAnchorResult:
-		fmt.Println(t.Anchor.ShortString())
+		out = t.Anchor.ShortString()
 	case goat.AsyncProbeResult:
-		fmt.Println(t.Probe.ShortString())
+		out = t.Probe.ShortString()
 	case goat.AsyncMeasurementResult:
-		fmt.Println(t.Measurement.ShortString())
+		out = t.Measurement.ShortString()
 	case goat.AsyncStatusCheckResult:
-		fmt.Println(t.Status.ShortString())
+		out = t.Status.ShortString()
 	default:
-		fmt.Printf("No output formatter defined for object type '%T'\n", t)
+		out = fmt.Sprintf("No output formatter defined for object type '%T'\n", t)
+	}
+
+	if out != "" {
+		fmt.Println(out)
+		total++
 	}
 }
 
 func finish() {
+	for _, last := range connectLastResults {
+		if last.Event == "connect" {
+			fmt.Println(connectTableEntry(
+				last.ProbeID,
+				last.Asn,
+				last.Prefix,
+				last.GetTimeStamp(),
+				time.Time{},
+				last.Controller,
+			))
+			total++
+		}
+	}
+
 	if verbose {
 		fmt.Printf("# %d results\n", total)
 	}
@@ -148,6 +177,14 @@ func SomeOutputNtp(res *result.NtpResult) string {
 }
 
 func SomeOutputConnection(res *result.ConnectionResult) string {
+	if connectTableOutput {
+		return someOutputConnectionTable(res)
+	} else {
+		return someOutputConnectionSimple(res)
+	}
+}
+
+func someOutputConnectionSimple(res *result.ConnectionResult) string {
 	var e string
 	switch res.Event {
 	case "connect":
@@ -157,7 +194,61 @@ func SomeOutputConnection(res *result.ConnectionResult) string {
 	default:
 		e = "?"
 	}
-	return res.BaseString() + fmt.Sprintf("\t%s\t%s", e, res.Controller)
+	sas := "N/A"
+	sprefix := "N/A"
+	if res.Asn != 0 {
+		sas = fmt.Sprintf("AS%d", res.Asn)
+		sprefix = res.Prefix.String()
+	}
+	return res.BaseString() + fmt.Sprintf("\t%s\t%s\t%s\t%s", e, sas, sprefix, res.Controller)
+}
+
+func someOutputConnectionTable(res *result.ConnectionResult) string {
+	ret := ""
+	if res.Event == "disconnect" {
+		from := time.Time{}
+		if connectLastResults[res.ProbeID] != nil {
+			from = connectLastResults[res.ProbeID].GetTimeStamp()
+		}
+		ret = connectTableEntry(
+			res.ProbeID,
+			res.Asn,
+			res.Prefix,
+			from,
+			time.Time(res.TimeStamp),
+			res.Controller,
+		)
+	}
+	connectLastResults[res.ProbeID] = res
+	return ret
+}
+
+func connectTableEntry(
+	probe uint,
+	as uint,
+	prefix netip.Prefix,
+	from time.Time,
+	until time.Time,
+	ctr string,
+) string {
+	asTime := func(t time.Time) string {
+		if t.IsZero() {
+			return "N/A"
+		} else {
+			return time.Time(t).UTC().Format("2006-01-02T15:04:05Z")
+		}
+	}
+	sas := "N/A"
+	sprefix := "N/A"
+	if as != 0 {
+		sas = fmt.Sprintf("AS%d", as)
+		sprefix = prefix.String()
+	}
+	dur := "N/A"
+	if !from.IsZero() && !until.IsZero() {
+		dur = fmt.Sprint(until.Sub(from))
+	}
+	return fmt.Sprintf("%d\t%s\t%s\t%v\t%v\t%s\t%v", probe, sas, sprefix, asTime(from), asTime(until), dur, ctr)
 }
 
 func SomeOutputUptime(res *result.UptimeResult) string {