+++ doc/changes.md
@@ -4,6 +4,7 @@
 
 * CHANGED: if the result download stop time is precisely on a day boundary, then it's adjusted
   to be 23:59:59 of the previous day instead
+* NEW: `some` output formatter for connections, including a processed `table` format
 
 ## v0.7.2